taking selectivity for non-indexed equi-join predicates
Made minor changes to the function table_multi_eq_cond_selectivity.

First the function was checking if there is a keyuse only then we were
trying to use the function to get estimates for the equi-join predicated.
I think this can calculate the estimates even if we don't have a keyuse.

Second change is done when we try to get the avg frequency of a column,
then we need to make sure that statistics are available in the statistical
tables. This check was missing and so get_column_avg_frequency() was
returning zero for columns.

Also fixed a typo.